separated into model and controller
