+++ phyloviz/model.py
@@ -6,7 +6,6 @@ import numpy as np
 import skbio
 import networkx as nx
 from Bio import Phylo
-from flask import Flask
 from scipy.cluster.hierarchy import complete
 
 from tornado.web import RequestHandler
@@ -46,7 +45,7 @@ def read(file_name, file_format='newick'):
         tree = skbio.read( file_name, file_format, into=TreeNode)
 
         return tree
-    
+
     return None
         # create node_metadata data frame
         #for node in tree.preorder():
@@ -686,47 +685,3 @@ class Model(object):
 
     def retrive_view_coords(self):
         return (self.node_metadata,self.edge_metadata)
-
-
-
-tree = read('./TreeOfLife.nwk','newick')
-m = Model(tree)
-nodeM, edgeM = m.retrive_view_coords()
-
-class IndexHandler(RequestHandler):
-    def get(self):
-        self.write({'hello':'world'})
-        self.finish()
-
-class ModelHandler(RequestHandler):
-    def get(self):
-        nodes = nodeM.to_json(orient='records')
-        edges = edgeM.to_json(orient='records')
-        self.render('tree.html', node_coords=nodes,
-                     edge_coords=edges)
-
-# # Set up REST API for model
-# app = Flask(__name__)
-
-# @app.route('/', methods=['GET'])
-# def hello_world():
-#     return "hello world!"
-
-# @app.route('/nodes', methods=['GET'])
-# def get_nodes():
-#     """ Returns node metadata dataframe as a json object
-#     with index orientation by default.
-#     """
-#     return m.node_metadata.to_json(orient='records')
-
-
-# @app.route('/edges', methods=['GET'])
-# def get_edges():
-#     """ Returns edge metadata dataframe as a json object
-#     with index orientation by default.
-#     """
-#     return m.edge_metadata.to_json(orient='records')
-
-# Run Flask app
-# if __name__ == '__main__':
-#     app.run(host=LOCALHOST, port=MODEL_PORT, debug=True)
